UT: Move and rename StubAgentLogRepository to InMemoryAgentLogRepository
